test(portable-text-editor): update tests to test for new normalization of marks and markdefs
